[step4] refactor: LastWinningLotto 테스트 리팩토링
- 지난 당첨번호, 보너스 번호에 대한 책임만 부여
- 당첨결과에 대한 책임은 일단 MyLottoTickets으로 이동
